Task 3 - Image Captioning with BLIP/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
Pillow>=9.0.0
requests>=2.28.0
gradio>=4.0.0
huggingface_hub>=0.19.0
